+++ modules/project/import-core/src/sap/commerce/toolset/project/configurator/ProjectEncodingConfigurator.kt
@@ -1,0 +1,40 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package sap.commerce.toolset.project.configurator
+
+import com.intellij.openapi.externalSystem.service.project.IdeModifiableModelsProvider
+import com.intellij.openapi.vfs.encoding.EncodingProjectManager
+import sap.commerce.toolset.project.descriptor.HybrisProjectDescriptor
+import java.nio.charset.StandardCharsets
+
+/**
+ * Let's assume that UTF-8 is the standard charset for all properties files.
+ * https://help.sap.com/docs/SAP_COMMERCE_CLOUD_PUBLIC_CLOUD/aa417173fe4a4ba5a473c93eb730a417/8b76856086691014a125e8e6d7d36f59.html#character-encoding
+ */
+class ProjectEncodingConfigurator : ProjectImportConfigurator {
+
+    override val name: String
+        get() = "Project Encoding"
+
+    override fun configure(hybrisProjectDescriptor: HybrisProjectDescriptor, modifiableModelsProvider: IdeModifiableModelsProvider) {
+        val project = hybrisProjectDescriptor.project ?: return
+
+        EncodingProjectManager.getInstance(project)
+            .setDefaultCharsetForPropertiesFiles(null, StandardCharsets.UTF_8)
+    }
+}