+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [2025.2.4.3]
 
+### `Project Import` enhancements
+- Auto-configure `UTF-8` as a charset for properties files [#1597](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1597)
+
 ### `CCv2` enhancements
 - Added `S-User` alias for service `modified by` characteristic [#1596](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1596)